% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPCP_mat.R
\name{OPCP_mat}
\alias{OPCP_mat}
\title{Calculate Observed Proportion of Concordant Pairs (OPCP)}
\usage{
OPCP_mat(data)
}
\arguments{
\item{data}{A data frame containing the variables specified in the formula.}
}
\value{
A matrix of OPCPs.
}
\description{
This function provides a matrix (like a correlation matrix) of Observed Proportions of Concordant Pairs (OPCPs) .
}
\examples{
# Example using the spi dataset from the psychTools package
library(psychTools)
sc <- psych::scoreVeryFast(spi.keys, spi)
spi_sc <- cbind(spi, sc)
spi_sc_age_sex_B5 <- spi_sc |>
  dplyr::select(age, sex, Agree, Consc, Neuro, Extra, Open)


OPCP_mat(data = spi_sc_age_sex_B5)

}
